# build(ui): improve performance with `esbuild-loader`