--- a/ui/package.json
+++ b/ui/package.json
@@ -86,6 +86,7 @@
         "yarn-deduplicate": "^6.0.2"
     },
     "resolutions": {
-        "semver": "^7.5.2"
+        "semver": "^7.5.2",
+        "types-ramda": "0.29.4"
     }
 }