--- a/ui/src/app/tsconfig.json
+++ b/ui/src/app/tsconfig.json
@@ -3,6 +3,7 @@
         "outDir": "./../../dist/app",
         "sourceMap": true,
         "noImplicitAny": true,
+        "isolatedModules": true, // for compatibility with esbuild
         "module": "ES2020", // must be ES2020+ for dynamic imports: https://github.com/Microsoft/TypeScript/issues/16675, https://github.com/webpack/webpack/issues/5703#issuecomment-357512412
         "moduleResolution": "node",
         "esModuleInterop": true,