<!-- markdownlint-disable MD041 -->

<!--

### Before you open your PR

- Run `make pre-commit -B` to fix codegen and lint problems (build will fail).
- [Signed-off your commits](https://github.com/apps/dco/) (otherwise the DCO check will fail).
- Used [a conventional commit message](https://www.conventionalcommits.org/en/v1.0.0/).

### When you open your PR

- PR title format should also conform to [conventional commits](https://www.conventionalcommits.org/en/v1.0.0/).
- "Fixes #" is in both the PR title (for release notes) and this description (to automatically link and close the issue).
- Create the PR as draft.
- Once builds are green, mark your PR "Ready for review".

When changes are requested, please address them and then dismiss the review to get it reviewed again.

-->

<!-- Does this PR fix an issue -->

Follow-up to #11628

### Motivation

<!-- TODO: Say why you made your changes. -->

- similar to [Argo CD](https://github.com/argoproj/argo-cd/blob/d9df2525c57d4870215a6ce149dbedd08ae05fdb/ui/src/app/webpack.config.js#L37) and many other projects, we can use `esbuild` instead of TS etc to significantly speed up build times
  - (as `esbuild` is written in Go and compiled instead of using interpreted TS/JS)
  - we don't use any Babel or TS plugins, so I think this codebase is a good candidate for that
  - it also simplifies the build process as well, removing & replacing many other deps
  - caveats:
    - `esbuild` does not do type-checking, so we have to run `tsc --noEmit` as part of the `lint` phase
    - `esbuild-loader` does not currently support React Fast Refresh (c.f. https://github.com/privatenumber/esbuild-loader/issues/250, https://github.com/pmmmwh/react-refresh-webpack-plugin/issues/366) <details>
      - `react-hot-loader` [was replaced](https://github.com/gaearon/react-hot-loader/tree/v4.13.1?tab=readme-ov-file#moving-towards-next-step) by [React Fast Refresh](https://github.com/pmmmwh/react-refresh-webpack-plugin) and is deprecated / hasn't received an update in 2 years
      - that being said, `esbuild` is significantly faster than using Babel or TS for builds, so hot reloading is nice but not necessary
        - also hot reloading can be buggy
        - plus, in my experience, the dev build often falls back to a full page refresh anyway

</details>

- resolve various warnings in the build, during the install, and in the console

### Modifications

<!-- TODO: Say what changes you made. -->
<!-- TODO: Attach screenshots if you changed the UI. -->

- remove `babel-loader`, `ts-loader`, `react-hot-loader`, and `source-map-loader` <details>
  - note that `source-map-loader` is still present in the `yarn.lock` as it is used under-the-hood by other deps
    - but the warnings from directly using `source-map-loader` are no longer present in the build
    - and the bundle size shrank a bit, seemingly bc there is no `autolinker` etc dep in the bundle, but potentially other reasons (I did not check in-depth)

</details>

- add `esbuild-loader`

- add `tsc --noEmit` to the `lint` script
- add [`isolatedModules`](https://www.typescriptlang.org/tsconfig#isolatedModules) config to [ensure compatibility](https://esbuild.github.io/content-types/#isolated-modules) with `esbuild`'s compiler
  - since it has no type-checking, it can't do full program analysis, meaning there are certain types it doesn't support
  - same as when [using Babel to compile TS](https://babeljs.io/docs/babel-plugin-transform-typescript#typescript-compiler-options)

- remove `mdl.hot` etc usage of `react-hot-loader`

#### Related config / performance changes

- further optimize dev builds by using cheaper `eval` sourcemaps as [recommended by Webpack for performance](https://webpack.js.org/configuration/devtool/)
  - prod builds should always have highest quality sourcemaps, so do not change those
  - this was also discussed in https://github.com/argoproj/argo-workflows/pull/11716#discussion_r1309661376

- remove unused `react-paginate` `exclude`
  - it's not a dep we currently use; it doesn't exist in [the lockfile](https://github.com/argoproj/argo-workflows/blob/fa27edb845d3e4407f27f4f383d3d78620e11ee3/ui/yarn.lock)
    - (nor in `argo-ui`'s [lockfile](https://github.com/argoproj/argo-ui/blob/c65a9520366b0c0cf0ac585618d029e60041a94d/yarn.lock), of which all prod deps would be in Workflows's lockfile as well)

### Verification

<!-- TODO: Say how you tested your changes. -->

1. `yarn build` works fine
1. `make start UI=true` works fine and UI seems to function as expected

#### Performance Testing

some simple performance testing when running natively on macOS, taken directly from Webpack's reported stats:
- before:
  - prod build: ~65s
  - dev build: ~22s
  - dev rebuild: ~1.8s
- after:
  - prod build: ~50s (~25% faster)
  - dev build: ~16s (~25% faster)
  - dev rebuild: ~.8s (~50% faster)

#### Future Work

these are not as good improvements as I would have liked, which mean there are bigger bottlenecks elsewhere
  - e.g. Webpack itself (analysis, source maps, running loaders), `sass-loader`, etc
    - `sass` has been slow in the past and is ripe to be replaced with more modern PostCSS as well
    - switching to `esbuild` directly may further improve these stats
      - i.e. instead of Webpack + `esbuild-loader`
  - independently building `argo-ui` instead of as part of the Workflows (and CD) builds per https://github.com/argoproj/argo-ui/issues/453
  - optimizing more deps per #12059 and related comments, such as removing `moment`, consolidating similar deps, etc
    

<!-- 
### Beyond this PR

Thank you for submitting this! Have you ever thought of becoming a Reviewer or Approver on the project? 

Argo Workflows is seeking more community involvement and ultimately more [Reviewers and Approvers](https://github.com/argoproj/argoproj/blob/main/community/membership.md) to help keep it viable. 
See [Sustainability Effort](https://github.com/argoproj/argo-workflows/blob/main/community/sustainability_effort.md) for more information. 

-->