@agilgur5 2024-01-14
Oh, interesting, due to the new `isolatedModules` config, the build is [failing](https://github.com/argoproj/argo-workflows/actions/runs/7516274931/job/20461171832?pr=12516) due to missing `import type`. i.e. this PR actually depends on #12514