@agilgur5 2024-02-03
Merged `main` since #12514 was merged, but getting a test/build [failure](https://github.com/argoproj/argo-workflows/actions/runs/7768757176/job/21187052202?pr=12516) in `argo-ui` now 😕 . I was trying to partly workaround this in #12158, but otherwise this requires a few changes upstream in `argo-ui` (where [my PRs](https://github.com/argoproj/argo-ui/pulls/agilgur5) have been open even longer 😕)

I guess I'll kick this back to draft as a result -- but uh may want my `argo-ui` PRs to be prioritized 😅