@juliev0 2024-02-03
> Merged `main` since #12514 was merged, but getting a [test failure](https://github.com/argoproj/argo-workflows/actions/runs/7768757176/job/21187052202?pr=12516) in `argo-ui` now 😕 . Interestingly enough, the build passes fine (which I had tested), it's only the tests' build that's failing, I think because the tooling is different (`ts-jest` uses the TS compiler directly, `esbuild-loader` uses `esbuild`)?
> 
> I was trying to partly workaround a CSS side-effect import in #12158 for similar reasons to this, but otherwise this requires a few changes upstream in `argo-ui` (where [my PRs](https://github.com/argoproj/argo-ui/pulls/agilgur5) have been open [even longer](https://cloud-native.slack.com/archives/C020XM04CUW/p1705097023721029) 😕)
> 
> I guess I'll kick this back to draft as a result since it is pretty blocked on `argo-ui` -- but uh may want my `argo-ui` PRs to be prioritized 😅 (the label does not exist there ofc)

Okay, yes!