@agilgur5 2024-05-11
I might be able to get this working again now that https://github.com/argoproj/argo-ui/pull/509 and https://github.com/argoproj/argo-ui/pull/552 have been merged