@agilgur5 2024-05-11
Guh, this is now [failing to typecheck](https://github.com/argoproj/argo-workflows/actions/runs/9045160674/job/24854615361?pr=12516) due to another dep, this time very transitive: https://github.com/ramda/ramda/issues/3415#issuecomment-1757260119, https://github.com/swagger-api/swagger-js/issues/3176, https://github.com/swagger-api/apidom/issues/3279