@agilgur5 2024-05-11
Ok had to fix that by pinning `types-ramda` as it requires TS v5 per  https://github.com/ramda/ramda/issues/3415#issuecomment-1758719123 and https://github.com/swagger-api/swagger-js/issues/3176#issuecomment-1759429014