@agilgur5 2024-05-11
> ### Verification
> 1. `yarn build` works fine
> 2. `make start UI=true` works fine and UI seems to function as expected

Re-verified, seems to all be fine